moved necessary code to server.js
